refactor(将env判断相关函数移动至单独的文件):
