+++ components/PWA.js
@@ -1,5 +1,6 @@
 import { compressImage } from '@/lib/notion/mapImage'
-import { isBrowser } from '../lib/utils'
+
+import { isBrowser } from '@/lib/utils/envUtils'
 
 /**
  * 初始化PWA
