+++ lib/lang.js
@@ -1,5 +1,5 @@
 import BLOG from '@/blog.config'
-import { isBrowser, mergeDeep } from '@/lib/utils'
+import { mergeDeep } from '@/lib/utils'
 import enUS from './lang/en-US'
 import frFR from './lang/fr-FR'
 import jaJP from './lang/ja-JP'
@@ -10,6 +10,7 @@ import zhTW from './lang/zh-TW'
 import { extractLangPrefix } from './utils/pageId'
 
 import { getQueryVariable } from '@/lib/utils/url'
+import { isBrowser } from '@/lib/utils/envUtils'
 
 /**
  * 在这里配置所有支持的语言
