+++ lib/notion/convertInnerUrl.js
@@ -1,6 +1,7 @@
 import { idToUuid } from 'notion-utils'
-import { checkStrIsNotionId, isBrowser } from '../utils'
+import { checkStrIsNotionId } from '../utils'
 import { getLastPartOfUrl } from '@/lib/utils/url'
+import { isBrowser } from '@/lib/utils/envUtils'
 
 /**
  * 处理页面内连接跳转:
