+++ lib/password.js
@@ -1,4 +1,4 @@
-import { isBrowser } from './utils'
+import { isBrowser } from '@/lib/utils/envUtils'
 
 /**
  * 获取默认密码
