+++ lib/utils/index.js
@@ -2,12 +2,6 @@
 import { memo } from 'react'
 
 /**
- * 判断是否客户端
- * @returns {boolean}
- */
-export const isBrowser = typeof window !== 'undefined'
-
-/**
  * 打乱数组
  * @param {*} array
  * @returns
@@ -23,13 +17,6 @@ export const shuffleArray = array => {
   return array
 }
 
-/**
- * google机器人
- * @returns
- */
-export const isSearchEngineBot =
-  typeof navigator !== 'undefined' &&
-  /Googlebot|bingbot|Baidu/.test(navigator.userAgent)
 /**
  * 组件持久化
  */
@@ -193,35 +180,6 @@ export function deepClone(obj) {
  * @returns
  */
 export const delay = ms => new Promise(resolve => setTimeout(resolve, ms))
-
-/**
- * 判断是否移动设备
- */
-export const isMobile = () => {
-  let isMobile = false
-  if (!isBrowser) {
-    return isMobile
-  }
-
-  // 这个判断会引发 TypeError: navigator.userAgentData.mobile is undefined 问题，导致博客无法正常工作
-  // if (!isMobile && navigator.userAgentData.mobile) {
-  //   isMobile = true
-  // }
-
-  if (!isMobile && /Mobi|Android|iPhone/i.test(navigator.userAgent)) {
-    isMobile = true
-  }
-
-  if (/Android|iPhone|iPad|iPod/i.test(navigator.platform)) {
-    isMobile = true
-  }
-
-  if (typeof window.orientation !== 'undefined') {
-    isMobile = true
-  }
-
-  return isMobile
-}
 
 /**
  * 扫描页面上的所有文本节点，将url格式的文本转为可点击链接
