+++ lib/utils/url.js
@@ -1,4 +1,5 @@
-import { checkStartWithHttp, isBrowser } from '@/lib/utils/index'
+import { checkStartWithHttp } from '@/lib/utils/index'
+import { isBrowser } from '@/lib/utils/envUtils'
 
 /**
  * 诸如 article/https://test.com 等被错误拼接前缀的slug进行处理
