+++ themes/fukasawa/components/BlogListPage.js
@@ -1,12 +1,13 @@
 import { AdSlot } from '@/components/GoogleAdsense'
 import { siteConfig } from '@/lib/config'
 import { useGlobal } from '@/lib/global'
-import { deepClone, isBrowser } from '@/lib/utils'
+import { deepClone } from '@/lib/utils'
 import { debounce } from 'lodash'
 import { useEffect, useState } from 'react'
 import BlogCard from './BlogCard'
 import BlogPostListEmpty from './BlogListEmpty'
 import PaginationSimple from './PaginationSimple'
+import { isBrowser } from '@/lib/utils/envUtils'
 /**
  * 文章列表分页表格
  * @param page 当前页
