+++ themes/theme.js
@@ -3,8 +3,8 @@ import * as ThemeComponents from '@theme-components'
 import getConfig from 'next/config'
 import dynamic from 'next/dynamic'
 import { useRouter } from 'next/router'
-import { isBrowser } from '../lib/utils'
 import { getQueryParam, getQueryVariable } from '@/lib/utils/url'
+import { isBrowser } from '@/lib/utils/envUtils'
 
 // 在next.config.js中扫描所有主题
 export const { THEMES = [] } = getConfig()?.publicRuntimeConfig || {}
